+++ media_kit_test/lib/common/globals.dart
@@ -1,3 +1,6 @@
 import 'package:flutter/foundation.dart';
+import 'package:media_kit_video/media_kit_video.dart';
 
-final enableHardwareAcceleration = ValueNotifier<bool>(true);
+final configuration = ValueNotifier<VideoControllerConfiguration>(
+  const VideoControllerConfiguration(enableHardwareAcceleration: true),
+);