+++ media_kit_test/lib/tests/02.single_player_multiple_video.dart
@@ -23,11 +23,11 @@ class _SinglePlayerMultipleVideoScreenState
   //       Here, two [VideoController]s are created for testing.
   late final VideoController controller0 = VideoController(
     player,
-    enableHardwareAcceleration: enableHardwareAcceleration.value,
+    configuration: configuration.value,
   );
   late final VideoController controller1 = VideoController(
     player,
-    enableHardwareAcceleration: enableHardwareAcceleration.value,
+    configuration: configuration.value,
   );
 
   @override