+++ media_kit_test/lib/tests/04.tabs_test.dart
@@ -57,7 +57,7 @@ class TabViewState extends State<TabView> {
   late final Player player = Player();
   late final VideoController controller = VideoController(
     player,
-    enableHardwareAcceleration: enableHardwareAcceleration.value,
+    configuration: configuration.value,
   );
 
   @override