+++ media_kit_video/lib/src/video_controller/platform_video_controller.dart
@@ -5,7 +5,10 @@
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 import 'dart:async';
 import 'package:flutter/widgets.dart';
+
 import 'package:media_kit/media_kit.dart';
+
+import 'package:media_kit_video/src/video_controller/video_controller.dart';
 
 /// {@template platform_video_controller}
 ///
@@ -22,14 +25,8 @@ abstract class PlatformVideoController {
   /// The [Player] instance associated with this instance.
   final Player player;
 
-  /// Fixed width of the video output.
-  int? width;
-
-  /// Fixed height of the video output.
-  int? height;
-
-  /// Whether hardware acceleration should be enabled or not.
-  final bool enableHardwareAcceleration;
+  /// User defined configuration for [VideoController].
+  final VideoControllerConfiguration configuration;
 
   /// Texture ID of the video output, registered with Flutter engine by the native implementation.
   final ValueNotifier<int?> id = ValueNotifier<int?>(null);
@@ -40,9 +37,7 @@ abstract class PlatformVideoController {
   /// {@macro platform_video_controller}
   PlatformVideoController(
     this.player,
-    this.width,
-    this.height,
-    this.enableHardwareAcceleration,
+    this.configuration,
   );
 
   /// Sets the required size of the video output.
@@ -55,16 +50,6 @@ abstract class PlatformVideoController {
     int? width,
     int? height,
   });
-
-  @override
-  String toString() => 'VideoController('
-      'player: $player, '
-      'width: $width, '
-      'height: $height, '
-      'enableHardwareAcceleration: $enableHardwareAcceleration, '
-      'id: $id, '
-      'rect: $rect'
-      ')';
 
   /// A [Future] that completes when the first video frame has been rendered.
   Future<void> get waitUntilFirstFrameRendered =>