+++ media_kit_video/linux/video_output.cc
@@ -26,7 +26,7 @@ struct _VideoOutput {
   mpv_render_context* render_context;
   gint64 width;
   gint64 height;
-  gboolean enable_hardware_acceleration;
+  VideoOutputConfiguration configuration;
   TextureUpdateCallback texture_update_callback;
   gpointer texture_update_callback_context;
   FlTextureRegistrar* texture_registrar;
@@ -72,7 +72,7 @@ static void video_output_init(VideoOutput* self) {
   self->render_context = NULL;
   self->width = 0;
   self->height = 0;
-  self->enable_hardware_acceleration = TRUE;
+  self->configuration = VideoOutputConfiguration{};
   self->texture_update_callback = NULL;
   self->texture_update_callback_context = NULL;
   self->texture_registrar = NULL;
@@ -83,45 +83,37 @@ static void video_output_init(VideoOutput* self) {
 VideoOutput* video_output_new(FlTextureRegistrar* texture_registrar,
                               FlView* view,
                               gint64 handle,
-                              gint64 width,
-                              gint64 height,
-                              gboolean enable_hardware_acceleration) {
+                              VideoOutputConfiguration configuration) {
   g_print("media_kit: VideoOutput: video_output_new: %ld\n", handle);
   VideoOutput* self = VIDEO_OUTPUT(g_object_new(video_output_get_type(), NULL));
   self->texture_registrar = texture_registrar;
   self->handle = (mpv_handle*)handle;
-  self->width = width;
-  self->height = height;
-#ifdef MPV_RENDER_API_TYPE_SW
-  self->enable_hardware_acceleration = enable_hardware_acceleration;
-#else
-  // Can't use S/W rendering if MPV_RENDER_API_TYPE_SW is not defined.
-  self->enable_hardware_acceleration = TRUE;
-  if (!enable_hardware_acceleration) {
+  self->width = configuration.width;
+  self->height = configuration.height;
+  self->configuration = configuration;
+#ifndef MPV_RENDER_API_TYPE_SW
+  // MPV_RENDER_API_TYPE_SW must be available for S/W rendering.
+  if (!self->configuration.enable_hardware_acceleration) {
     g_printerr("media_kit: VideoOutput: S/W rendering is not supported.\n");
   }
+  self->configuration.enable_hardware_acceleration = TRUE;
 #endif
-  mpv_set_option_string(self->handle, "vo", "libmpv");
   mpv_set_option_string(self->handle, "video-sync", "audio");
   mpv_set_option_string(self->handle, "video-timing-offset", "0");
   gboolean hardware_acceleration_supported = FALSE;
-  if (self->enable_hardware_acceleration) {
+  if (self->configuration.enable_hardware_acceleration) {
     GError* error = NULL;
     GdkWindow* window = gtk_widget_get_window(GTK_WIDGET(view));
-    self->gdk_gl_context =
-        gdk_window_create_gl_context(window, &error);
+    self->gdk_gl_context = gdk_window_create_gl_context(window, &error);
     if (error == NULL) {
       // OpenGL context must be made current before creating mpv render context.
       gdk_gl_context_realize(self->gdk_gl_context, &error);
       if (error == NULL) {
         // Create |FlTextureGL| and register it.
         self->texture_gl = texture_gl_new(self);
-        gboolean result = fl_texture_registrar_register_texture(
-            texture_registrar, FL_TEXTURE(self->texture_gl));
-        g_print("fl_texture_registrar_register_texture: %d\n", result);
-        if (result) {
+        if (fl_texture_registrar_register_texture(
+                texture_registrar, FL_TEXTURE(self->texture_gl))) {
           // Request H/W decoding.
-          mpv_set_option_string(self->handle, "hwdec", "auto");
           mpv_opengl_init_params gl_init_params{
               [](auto, auto name) {
                 GdkDisplay* display = gdk_display_get_default();
@@ -151,10 +143,8 @@ VideoOutput* video_output_new(FlTextureRegistrar* texture_registrar,
             params[2].type = MPV_RENDER_PARAM_X11_DISPLAY;
             params[2].data = gdk_x11_display_get_xdisplay(display);
           }
-          gint result = mpv_render_context_create(&self->render_context,
-                                                  self->handle, params);
-          g_print("mpv_render_context_create: %d\n", result);
-          if (result == 0) {
+          if (mpv_render_context_create(&self->render_context, self->handle,
+                                        params) == 0) {
             mpv_render_context_set_update_callback(
                 self->render_context,
                 [](void* data) {